refactoring

+++ module3/src/main/java/nix/command/GetSubjectBestAndWorstResults.java
@@ -2,6 +2,7 @@ package nix.command;
 
 import nix.service.UniversityService;
 
+import java.math.BigDecimal;
 import java.util.Map;
 
 public class GetSubjectBestAndWorstResults implements Command {
@@ -9,7 +10,7 @@ public class GetSubjectBestAndWorstResults implements Command {
 
     @Override
     public void execute() {
-        for (Map.Entry<String, Double> entry : UNIVERSITY_SERVICE.getSubjectBestAndWorstResults().entrySet()) {
+        for (Map.Entry<String, BigDecimal> entry : UNIVERSITY_SERVICE.getSubjectBestAndWorstResults().entrySet()) {
             System.out.println("Key: " + entry.getKey() + " Value: " + entry.getValue());
         }
     }
